+++ src/app/dashboard/common/status.types.ts
@@ -1,6 +1,6 @@
 export type ComputationState = 'PENDING' | 'STARTED' | 'SUCCESS' | 'FAILURE' | 'RETRY' | 'REVOKED'
 
-export type ComputationFlag = 'NEW' | 'ARCHIVED' | null
+export type ComputationFlags = ('NEW' | 'ARCHIVED' | 'DEMO')[]
 
 export type ComputationStateInfo = {
     state: ComputationState